+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 🛠 What's changed
-- Add change, read, uniform, add, remove, register, apply, tween, property for Lua style
-- Add sub for other styles
+- Add template, step and reset
+- Add Parenthesis Validator